Uploading diaObject duplication issue notebook
Uploading notebook that demonstrates and characterizes the DiaObject duplication issue identified in notebook DP02_07b "Variable Stars in DP0.2," where the same variable star can be identified with multiple DiaOjbectIds. The key goal of this notebook is to inform how this issue may or may not affect your scientific analysis of transients and/or variables.